Add config file, logs retrieval, logs server pages
* Adds the ability to provide a configuration file which specifies
which clients should be run. Currently only used for specifying
if log files should be sent

* Creates a new LogInfo class which is used to retrieve log files
based on the configuration and send them back to the server. This
will retrieve all the log files specified in the config, although
this should be made more resistant to failures or missing files

* Adds a log page to the server which will show all the log files
that have been sent by the client. These should probably be broken
up into individual pages

* Adds a main menu append function which adds a main menu reference
which jumps back to the main page to each page that is returned by
the server